ETSWORK-16463 remove typescript from custom component example
