+++ examples/react/json-based-microapp-renderer-with-css/src/styles.css
@@ -13,7 +13,7 @@
     0px 3px 4px 0px rgba(0, 0, 0, 0.14), 0px 1px 8px 0px rgba(0, 0, 0, 0.12);
   padding: 5px;
   margin-top: 15px;
-  width: 100%;
+  width: 400px;
   max-width: 400px;
 }
 
@@ -87,10 +87,10 @@ button.ac-pushButton:is(:hover) {
     box-shadow 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,
     border-color 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,
     color 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
-  color: blue;
+  color: purple;
   text-decoration: none;
   background-color: rgba(39, 133, 176, 0.04);
-  border: 1px solid blue;
+  border: 1px solid purple;
 }
 
 select {
@@ -144,8 +144,6 @@ select {
   justify-content: center;
   padding-top: 50px;
   padding-bottom: 50px;
-  /* display: inline-block;
-  position: relative; */
   max-width: 400px;
   height: 80px;
 }
