+++ examples/react/json-based-microapp-renderer-with-custom-components/src/components/Loader.jsx
