--- examples/react/json-based-microapp-renderer-with-custom-components/src/styles.css
@@ -1,4 +1,0 @@
-.App {
-  font-family: sans-serif;
-  text-align: center;
-}
